# api_glewe/api_glewe/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " OpenApi Documentation",
        "description": " Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "operationId": "01ffcf3e5dd2e813d962b3836b302fb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "last_name": {
                                        "type": "string",
                                        "example": "hanane"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "celsia"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "hananecelsia"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "hananecelsia@gmail.com"
                                    },
                                    "country_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success register"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "operationId": "8ab039c9fd5bf60c74e218987419dda8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "hananecelsia"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "hananecelsia@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success register"
                    }
                }
            }
        },
        "/api/params/new-category": {
            "post": {
                "tags": [
                    "new category"
                ],
                "operationId": "86a9ec6dff29fbd37c61247b9127a5b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "hanane"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "category created"
                    }
                }
            }
        },
...
    }
}